refactor Transaction Item
